refactor stereo related code a lot! (not yet functional)
