+++ src/machine_vision_acquisition_python/calibration/libcalib.py
@@ -5,25 +5,11 @@ import json
 import logging
 from pathlib import Path
 from typing import List
-
+from machine_vision_acquisition_python.calibration.shared import Calibration
 import numpy as np
-from numpy.typing import NDArray
+
 
 log = logging.getLogger(__name__)
-
-
-TEMP_FILE_PATH = r"/mnt/powerplant/input/projects/dhs/smartsensingandimaging/development/fops/2022-04-29/calibration-images/caloutput.json"
-
-
-class Calibration:
-    def __init__(self, name_or_serial, cameraMatrix, distCoeffs, rvec, tvec, image_width, image_height) -> None:
-        self.serial: str = name_or_serial
-        self.cameraMatrix: NDArray = cameraMatrix
-        self.distCoeffs: NDArray = distCoeffs
-        self.rvec: NDArray = rvec
-        self.tvec: NDArray = tvec
-        self.image_width: int = image_width
-        self.image_height: int = image_height
 
 
 def read_calib_parameters(calibio_json: Path):
